ai_portal/requirements.txt
# AI Portal - Python Dependencies
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# HTTP clients and async
aiohttp==3.9.1
httpx==0.25.2
asyncio-timeout==4.0.3

# AI and ML services
openai==1.3.7
google-cloud-aiplatform==1.38.1
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Memory and learning system
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
torch==2.1.1
transformers==4.36.2

# Web automation and scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30
pathlib2==2.3.7
typing-extensions==4.8.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production server (optional)
gunicorn==21.2.0